refactor stylus injector
