+++ source/css/main.styl
@@ -1,13 +1,13 @@
 // Variables Layer
 // --------------------------------------------------
 @import "_variables/base";
-for $inject_variable in hexo-config('injects.variable')
+for $inject_variable in injector('variable')
   @import $inject_variable;
 
 // Mixins Layer
 // --------------------------------------------------
 @import "_mixins/base";
-for $inject_mixin in hexo-config('injects.mixin')
+for $inject_mixin in injector('mixin')
   @import $inject_mixin;
 
 // Common Layer
@@ -16,5 +16,5 @@ for $inject_mixin in hexo-config('injects.mixin')
 
 // Custom Layer
 // --------------------------------------------------
-for $inject_style in hexo-config('injects.style')
+for $inject_style in injector('style')
   @import $inject_style;
